backend/pyproject.toml
[tool.poetry]
name = "ecoeats-backend"
version = "0.1.0"
description = "EcoEats FastAPI backend with recipe recommendations"
authors = ["EcoEats Team <team@ecoeats.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"
alembic = "^1.14.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"

[tool.fly]
app = "ecoeats-backend-vkqkjjcg"
primary_region = "iad"

[tool.fly.build]
builder = "paketobuildpacks/builder:base"

[tool.fly.env]
PORT = "8000"

[[tool.fly.services]]
http_checks = []
internal_port = 8000
processes = ["app"]
protocol = "tcp"
script_checks = []

  [tool.fly.services.concurrency]
  hard_limit = 25
  soft_limit = 20
  type = "connections"

  [[tool.fly.services.ports]]
  force_https = true
  handlers = ["http"]
  port = 80

  [[tool.fly.services.ports]]
  handlers = ["tls", "http"]
  port = 443

  [[tool.fly.services.tcp_checks]]
  grace_period = "1s"
  interval = "15s"
  restart_limit = 0
  timeout = "2s"
